// models/eleve.ts
export interface Eleve {
  id:number;
  nom: string;
  prenom: string;
  civilite: string;
  penseBete :string;
  adressePostale: string;
  lieuN: string;
  signiature: string;
  creePar:number;
  modifiePar:number;
  createdAt: string;
  updatedAt: string;
  genre: string;
  idCU:number;
  idRole:number;
  nomE:string;
  nomC:string;
  abrege:string;
  idEcole:string;

}